Mon, Jan 16, 2023, 12:09 PM +00:00

+++ runoff/try.c
@@ -7,6 +7,6 @@ int candidates_count=10;
 
 int main(void)
 {
-    double percent = candidates_votes / candidates_count;
+    float percent = (float) candidates_votes / candidates_count;
     printf("%f \n",percent);
 }
